+++ src/main/java/de/mossgrabers/convertwithmoss/format/kmp/KMPCreator.java
@@ -116,7 +116,7 @@ public class KMPCreator extends AbstractCreator<KMPCreatorUI>
     {
         if (detectSettings.reduceBitDepth <= 0 || SUPPORTED_BIT_DEPTHS.contains (Integer.valueOf (detectSettings.reduceBitDepth)))
             return true;
-        this.notifier.log ("IDS_PROCESSING_REDUCE_BITE_DEPTH_NOT_SUPPORTED", Integer.toString (detectSettings.reduceBitDepth), "8, 16");
+        this.notifier.log ("IDS_PROCESSING_REDUCE_BIT_DEPTH_NOT_SUPPORTED", Integer.toString (detectSettings.reduceBitDepth), "8, 16");
         return false;
     }