+++ src/main/java/de/mossgrabers/convertwithmoss/format/sf2/Sf2Creator.java
@@ -114,7 +114,7 @@ public class Sf2Creator extends AbstractCreator<Sf2CreatorUI>
     {
         if (detectSettings.reduceBitDepth <= 0 || SUPPORTED_BIT_DEPTHS.contains (Integer.valueOf (detectSettings.reduceBitDepth)))
             return true;
-        this.notifier.log ("IDS_PROCESSING_REDUCE_BITE_DEPTH_NOT_SUPPORTED", Integer.toString (detectSettings.reduceBitDepth), "16, 24");
+        this.notifier.log ("IDS_PROCESSING_REDUCE_BIT_DEPTH_NOT_SUPPORTED", Integer.toString (detectSettings.reduceBitDepth), "16, 24");
         return false;
     }